requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.1.2
gunicorn==23.0.0
numpy==1.21.0
python-dotenv==0.19.0
requests==2.26.0
flask-cors==3.0.10
openai==0.27.0
PyJWT==2.1.0
Flask-Session==0.4.0
gql==3.4.1
redis==5.0.1
flask-caching==2.1.0
pandas==2.1.1
matplotlib==3.8.0
scikit-learn==1.4.0
